% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\docType{package}
\name{astror}
\alias{astror}
\alias{astror-package}
\title{astror is a package for scaling fits images into a range displayable by a
standard monitor.}
\description{
astror is a package for scaling fits images into a range displayable by a
standard monitor.
}

