% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctools.r
\name{bin.lc}
\alias{bin.lc}
\title{Bin Light Curve}
\usage{
bin.lc(lc, bin.width)
}
\arguments{
\item{lc}{Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{bin.width}{Time delta width of the time bins, must be in the same units as lc$TIME}
}
\value{
Light curve data frame binned into evenly spaced time bins with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE
}
\description{
Bins a light curve into evenly spaced time bins
}
\examples{
\dontrun{
lightcurve <- bin.lc(lightcurve, 100)
}
}
\author{
Derek Blue
}
