% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{clampAbovePercentile}
\alias{clampAbovePercentile}
\title{Clamp the highest values to a percentile}
\usage{
clampAbovePercentile(values, fraction = 0.99)
}
\arguments{
\item{values}{the values to remove the top tail from}

\item{fraction}{the fraction of values to clamp down}
}
\description{
Clamp the highest values to a percentile
}

