% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{clampBelowPercentile}
\alias{clampBelowPercentile}
\title{Clamp the lowest values to a percentile}
\usage{
clampBelowPercentile(values, fraction = 0.01)
}
\arguments{
\item{values}{the values to remove the bottom tail from}

\item{fraction}{the fraction of values to clamp up}
}
\description{
Clamp the lowest values to a percentile
}

