% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{clampPercentile}
\alias{clampPercentile}
\title{Clamp the lowest and highest values.}
\usage{
clampPercentile(values, below = 0.01, above = 0.99)
}
\arguments{
\item{values}{The values to remove the top and bottom from}

\item{below}{Percentile below which to clamp up}

\item{above}{Percentile above which to camp down}
}
\description{
Clamp the lowest and highest values.
}

