% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.r
\name{correct.area}
\alias{correct.area}
\title{Correct area}
\usage{
correct.area(spectrum, arf.path, energy.bins)
}
\arguments{
\item{spectrum}{Absolute or relative file path to the spectrum file}

\item{arf_path}{Absolute or relative file path to the ARF file}

\item{energy_bins}{Number of energy bins}
}
\value{
Fluxed spectrum data
}
\description{
From ARF?
}
\examples{
\dontrun{
lc <- xmm.lc("obsid_pn_lccor_300-10000.fits")
}
}
\author{
Derek Blue
}
