% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{dcf.plot}
\alias{dcf.plot}
\title{Plot DCF}
\usage{
dcf.plot(
  dcf.df,
  contours = FALSE,
  xlab = "tau [days]",
  ylab = "dcf",
  plt.title = "Broadband UV DCF"
)
}
\arguments{
\item{dcf.df}{DCF data frame with strcture: tau, dcf, p90, p95, p99, n90, n95, n99}

\item{contours}{optional: boolean to include or exclude confidence contours}

\item{ylab}{optional: Y axis label}

\item{plt.title}{optional: Plot title}
}
\value{
Plot of the DCF
}
\description{
Plots the DCF
}
\examples{
\dontrun{
plt <- dcf.plot(dcf.df)
plot(plt)
}
}
\author{
Derek Blue
}
