% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dcf.r
\name{dcf}
\alias{dcf}
\title{Run Discrete Correlation Function}
\usage{
dcf(base, compare, beta, bin.width = 1, nsims = 10)
}
\arguments{
\item{base}{required data frame: Light curve data frame to use as base for DCF. Expects structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{compare}{required data frame: Light curve data frame to compare to base for DCF. Expects structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{beta}{required numeric: PSD slope to use when simmulating light curves, should be the PSD slope of compare}

\item{bin.width}{optional numeric: DCF bin widths for binning of the final DCF}

\item{nsims}{optional numeric: number of simulations to run for DCF confidence contours}
}
\value{
DCF data frame with strcture: tau, dcf, p90, p95, p99, n90, n95, n99
}
\description{
Runs a discrete correlation function on the input light curves following Edelson & Krolik 1988, using https://github.com/svdataman/sour
}
\examples{
\dontrun{
xmm.lc <- xmm.pn.lc(xmm.pn.file.path)
xmm.uv.lc <- xmm.uvw1.lc(xmm.uvw1.file.path)
dcf <- run.dcf(xmm.uv.lc, xmm.lc, beta, bin.width = 1, nsims = 10)
}
}
\author{
Derek Blue
}
