% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{extractLuminance}
\alias{extractLuminance}
\title{This is a function to extract log luminance values from the a set of RGB
channels for use with the gradient domain high dynamic range compression
algorithm.}
\usage{
extractLuminance(red, green, blue)
}
\arguments{
\item{red}{The red channel}

\item{green}{The green channel}

\item{blue}{The blue channel}
}
\description{
This is a function to extract log luminance values from the a set of RGB
channels for use with the gradient domain high dynamic range compression
algorithm.
}

