% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{ff.plot}
\alias{ff.plot}
\title{Plot Flux Flux}
\usage{
ff.plot(
  ff.df,
  xlab = "Soft Rate [count/s]",
  ylab = "Hard Rate [count/s]",
  plt.title = "Flux-Flux Plot"
)
}
\arguments{
\item{ff.df}{required data frame: Flux flux data frame with structure: SOFT.RATE, SOFT.ERROR, HARD.RATE, HARD.ERROR}

\item{xlab}{optional string: X axis label}

\item{ylab}{optional string: Y axis label}

\item{plt.title}{optional string: Plot title}
}
\value{
Plot of the hard band flux vs soft band flux
}
\description{
Creates a plot of the hard band flux vs soft band flux
}
\examples{
\dontrun{
plt <- ff.plot(ff.df)
plot(plt)
}
}
\author{
Derek Blue
}
