% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctools.r
\name{fvar.edelson}
\alias{fvar.edelson}
\title{Fractional Variability (Edelson)}
\usage{
fvar.edelson(lc)
}
\arguments{
\item{lc}{Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}
}
\value{
List containing fractional variability and error for the light curve
}
\description{
Calculates the fractional variability following Edelson et al. 2002, ApJ, 568, 610
}
\examples{
\dontrun{
fv <- fvar.edelson(lightcurve)
fvar <- fv[1]
fvar.err <- fv[2]
}
}
\author{
Derek Blue
}
