% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctools.r
\name{fvar.ponti}
\alias{fvar.ponti}
\title{Fractional Variability (Ponti)}
\usage{
fvar.ponti(lc)
}
\arguments{
\item{lc}{Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}
}
\value{
List containing fractional variability and error for the light curve
}
\description{
Calculates the fractional variability following Ponti et al. 2004, A&A, 417, 451
}
\examples{
\dontrun{
fv <- fvar.ponti(lightcurve)
fvar <- fv[1]
fvar.err <- fv[2]
}
}
\author{
Derek Blue
}
