% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctools.r
\name{fvar.vaughan}
\alias{fvar.vaughan}
\title{Fractional Variability (Vaughan)}
\usage{
fvar.vaughan(lc)
}
\arguments{
\item{lc}{Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}
}
\value{
List containing fractional variability and error for the light curve
}
\description{
Calculates the fractional variability following Vaughan et al. 2003, MNRAS, 345, 1271
}
\examples{
\dontrun{
fv <- fvar.vaughan(lightcurve)
fvar <- fv[1]
fvar.err <- fv[2]
}
}
\author{
Derek Blue
}
