% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{geometricScale}
\alias{geometricScale}
\title{Geometrically scale a vector or matrix of values to fit in the range [0,1]}
\usage{
geometricScale(values)
}
\arguments{
\item{values}{the values [vector or matrix] to scale}
}
\description{
Applies a geometric scaling function to ensure that all values are within
the range [0,1]. The smallest value will always be mapped to 0, and the largest
to 1.
}

