% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctools.r
\name{hard.ratio}
\alias{hard.ratio}
\title{Hardness Ratio}
\usage{
hard.ratio(slc, hlc)
}
\arguments{
\item{slc}{required data frame: 0.3 - 1 keV light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{hlc}{required data frame: 2 - 10 keV light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}
}
\value{
Hardness ratio data frame with structure: TIME, TIMED, RATIO, ERROR
}
\description{
Calculates the hardness ratio
}
\examples{
\dontrun{
hrat <- bin.lc(soft.lc, hard.lc)
}
}
\author{
Derek Blue
}
