% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{hr.plot}
\alias{hr.plot}
\title{Plot Hardness Ratio}
\usage{
hr.plot(
  hr.df,
  xlab = "Time [s]",
  ylab = "Ratio",
  plt.title = "Hardness Ratio",
  show.origin = FALSE
)
}
\arguments{
\item{hr.df}{Hardness Ratio data frame with structure: TIME, RATIO, ERROR}

\item{xlab}{optional: X axis label}

\item{ylab}{optional: Y axis label}

\item{plt.title}{optional: Plot title}
}
\value{
Plot of the source and background light curve
}
\description{
Plots the source and background of a light curve
}
\examples{
\dontrun{
plt <- hr.plot(lightcurve)
plot(plt)
}
}
\author{
Derek Blue
}
