% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{lc.plot}
\alias{lc.plot}
\title{Plot Light Curve}
\usage{
lc.plot(lc, xlab = "Time [s]", ylab = "Rate [count/s]")
}
\arguments{
\item{lc}{Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{xlab}{optional: X axis label}

\item{ylab}{optional: Y axis label}
}
\value{
Plot of the source and background light curve
}
\description{
Plots the source and background of a light curve
}
\examples{
\dontrun{
plt <- lc.plot(lightcurve)
plot(plt)
}
}
\author{
Derek Blue
}
