% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{linearScale}
\alias{linearScale}
\title{Linearly scale values to fit in the range 0-1.}
\usage{
linearScale(values)
}
\arguments{
\item{values}{the values [matrix or vector] to scale}
}
\description{
This applies a linear scale to all of the values in the range such that
after the transformation all values are in the range [0,1].
}

