% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{logLuminance}
\alias{logLuminance}
\title{Function to safely log the lumininance values for processing}
\usage{
logLuminance(values)
}
\arguments{
\item{values}{vector of luminence values to compress for processing}
}
\description{
Function to safely log the lumininance values for processing
}

