% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{luminanceToColour}
\alias{luminanceToColour}
\title{Function to apply the new luminance values to the original image.}
\usage{
luminanceToColour(red, green, blue, luminance, saturation = 1)
}
\arguments{
\item{red}{The red channel}

\item{green}{The green channel}

\item{blue}{The blue channel}

\item{luminance}{The new luminance values}

\item{Saturation}{constant (must be > 0)}
}
\description{
Function to apply the new luminance values to the original image.
}

