% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{luminanceToGray}
\alias{luminanceToGray}
\title{Convert modified luminance values to the range [0,1] so they can be used to
actually draw an image.}
\usage{
luminanceToGray(luminance)
}
\arguments{
\item{luminance}{luminance values to re-exponentiate and make valid for
use in image creation.}
}
\description{
Convert modified luminance values to the range [0,1] so they can be used to
actually draw an image.
}

