% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctools.r
\name{prep.ff}
\alias{prep.ff}
\title{Prep Flux Flux}
\usage{
prep.ff(slc, hlc)
}
\arguments{
\item{slc}{0.3 - 1 keV light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{hlc}{2 - 10 keV light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}
}
\value{
Flux flux data frame with structre: SOFT.RATE, SOFT.ERROR, HARD.RATE, HARD.ERROR
}
\description{
Creates a flux flux data frame from the hard and soft band light curves
}
\examples{
\dontrun{
ff.df <- prep.ff(soft.lc, hard.lc)
}
}
\author{
Derek Blue
}
