% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{save.plot}
\alias{save.plot}
\title{Save Plot}
\usage{
save.plot(filepath, image.width = 7.24, image.height = 3.46)
}
\arguments{
\item{filepath}{required string: Absolute or relative file path for the saved plot}

\item{image.width}{optional numeric: Width of plot, defaults to 7.24}

\item{image.height}{optional numeric: Height of plot, defaults to 3.46}
}
\description{
Saves a plot to a png image
}
\examples{
\dontrun{
plt <- dcf.plot(dcf.df)
plot(plt)
}
}
\author{
Derek Blue
}
