% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctools.r
\name{sim.lc.lag}
\alias{sim.lc.lag}
\title{Simulate Lagged Light Curves}
\usage{
sim.lc.lag(
  beta,
  bins = 1024,
  length = 1e+05,
  time.lag = 0,
  freq.range = c(1e-04, 0.001),
  scale.factor = 1,
  shift.factor = 0
)
}
\arguments{
\item{beta}{Slope of the PSD for the simulated light curve}

\item{bins}{Number of data points for the simulated light curve, defaults to 1024}

\item{length}{Observation length, in seconds, for the simulated light curve, defaults to 100000}

\item{time.lag}{Lag in seconds to apply to the second light curve}

\item{freq.range}{Frequency range in Fourier space to apply the time lag}

\item{scale.factor}{Scaling factor for the simulated light curves, defaults to 1}

\item{shift.factor}{Shift factor for the simulated light curves, defaults to 0}
}
\value{
Simulated light curve data frames with structure: TIME, TIMED, RATE
}
\description{
Simulates two light curve from a given power spectrum with a time lag
}
\examples{
\dontrun{
lcdf <- sim.lc.lag(lightcurve, 100, time.lag = 114)
lc1 <- lcdf[[1]]
lc2 <- lcdf[[2]]
}
}
\author{
Derek Blue
}
