% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scales.R
\name{slog}
\alias{slog}
\title{Log scale the values.}
\usage{
slog(values)
}
\arguments{
\item{values}{the values [or matrix of values] to log scale}
}
\description{
The applies a log scale to all the values after shifting them all up [or
down] such that they the lowest value is about 1.01. This ensures that the
logarithm behaves in a sensible fashion, and can be applied repeatedly.
}

