% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf.r
\name{structure.function}
\alias{structure.function}
\title{Structure Function}
\usage{
structure.function(lc)
}
\arguments{
\item{lc}{Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}
}
\value{
List containing structure function and error for the light curve
}
\description{
Calculates the structure funcction
}
\examples{
\dontrun{
sfdf <- structure.function(lightcurve)
sf <- sfdf[1]
sf.err <- sfdf[2]
}
}
\author{
Derek Blue
}
