% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suzaku.r
\name{suzaku.lc}
\alias{suzaku.lc}
\title{Load Suzaku Light Curve}
\usage{
suzaku.lc(scpath, bgpath)
}
\arguments{
\item{scpath}{required string: Absolute or relative file path to the source light curve}

\item{bgpath}{required string: Absolute or relative file path to the background light curve}
}
\value{
Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE
}
\description{
Loads Suzaku light curves created using xselect, requires both source and background light curves
}
\examples{
\dontrun{
lc <- suzaku.lc("obsid_xi03_300-10000_sc.fits", "obsid_xi03_300-10000_bg.fits")
}
}
\author{
Derek Blue
}
