% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.r
\name{swift.bmag2cr}
\alias{swift.bmag2cr}
\title{Swift B Magnitude to Count Rate}
\usage{
swift.bmag2cr(mag)
}
\arguments{
\item{mag}{Swift B magnitude}
}
\value{
Swift B count rate
}
\description{
Converts the Swift B Magnitude to count rate
}
\examples{
\dontrun{
swift.b$RATE <- swift.bmag2cr(swift.b$MAG)
}
}
\author{
Derek Blue
}
