% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swift.r
\name{swift.table}
\alias{swift.table}
\title{Load Swift data table}
\usage{
swift.table(filepath, skip.rows = 0)
}
\arguments{
\item{filepath}{required string: Absolute or relative file path to the data file}

\item{skip.rows}{optional numeric: Number of rows to skip in the data file, defaults to 0}
}
\value{
Light curve data frame with structure: MJD, CR_X, X.ERR, HR, HR.ERR, V.MAG, V.ERR, B.MAG, B.ERR, U.MAG, U.ERR, W1.MAG, W1.ERR, M2.MAG, M2.ERR, W2.MAG, W2.ERR
}
\description{
Loads Swift data from text file
}
\examples{
\dontrun{
lc <- xmm.pn.lc("obsid_pn_lccor_300-10000.fits")
}
}
\author{
Derek Blue
}
