% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.r
\name{swift.w2mag2cr}
\alias{swift.w2mag2cr}
\title{Swift UVW2 Magnitude to Count Rate}
\usage{
swift.w2mag2cr(mag)
}
\arguments{
\item{mag}{required numeric: Swift UVW2 magnitude}
}
\value{
Swift UVW2 count rate
}
\description{
Converts the Swift UVW2 Magnitude to count rate
}
\examples{
\dontrun{
swift.uvw2$RATE <- swift.w2mag2cr(swift.uvw2$MAG)
}
}
\author{
Derek Blue
}
