% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmm.r
\name{xmm.arf}
\alias{xmm.arf}
\title{Load XMM Newton EPIC PN ARF}
\usage{
xmm.arf(filepath)
}
\arguments{
\item{filepath}{Absolute or relative file path to the "*.arf" file}
}
\value{
ARF data frame with structure: ENERGY_LO ENERGY_HI SPECRESP
}
\description{
Loads XMM Newton EPIC PN ARF file generated by the XMM SAS pipeline
}
\examples{
\dontrun{
arf <- xmm.arf("xmm_pn.arf")
}
}
\author{
Derek Blue
}
