% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmm.r
\name{xmm.pn.lc}
\alias{xmm.pn.lc}
\title{Load XMM Newton EPIC PN Light Curve}
\usage{
xmm.pn.lc(filepath)
}
\arguments{
\item{filepath}{required string: Absolute or relative file path to the "*lccorr*.fits" file}
}
\value{
Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE
}
\description{
Loads XMM Newton EPIC PN corrected light curves generated by the XMM SAS epiclccorr task
}
\examples{
\dontrun{
lc <- xmm.pn.lc("obsid_pn_lccor_300-10000.fits")
}
}
\author{
Derek Blue
}
