% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmm.r
\name{xmm.rmf}
\alias{xmm.rmf}
\title{Load XMM Newton EPIC PN RMF}
\usage{
xmm.rmf(filepath)
}
\arguments{
\item{filepath}{Absolute or relative file path to the "*.rmf" file}
}
\value{
ARF data frame with structure: ENERGY_LO ENERGY_HI SPECRESP
}
\description{
Loads XMM Newton EPIC PN RMF file generated by the XMM SAS pipeline
}
\examples{
\dontrun{
rmf <- xmm.rmf("xmm_pn.arf")
}
}
\author{
Derek Blue
}
