% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xmm.r
\name{xmm.uvw1.lc}
\alias{xmm.uvw1.lc}
\title{Load XMM Newton UVW1 Light Curve}
\usage{
xmm.uvw1.lc(filepath)
}
\arguments{
\item{filepath}{Absolute or relative file path to the "*lccorr*.fits" file}
}
\value{
Light curve data frame with structure: TIME, TIMED, RATE
}
\description{
Loads XMM Newton UVW1 light curves
}
\examples{
\dontrun{
lc <- xmm.lc("obsid_pn_lccor_300-10000.fits")
}
}
\author{
Derek Blue
}
