% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{dcf.plot}
\alias{dcf.plot}
\title{Plot DCF}
\usage{
dcf.plot(
  dcf.df,
  contours = FALSE,
  xlab = "tau [days]",
  ylab = "dcf",
  plt.title = "Broadband UV DCF"
)
}
\arguments{
\item{dcf.df}{required data frame: DCF data frame with strcture: tau, dcf, p90, p95, p99, n90, n95, n99}

\item{contours}{optional boolean: If TRUE plot the 90\%, 95\%, and 99.99\% confidence contours, defaults to FALSE}

\item{xlab}{optional string: X axis label}

\item{ylab}{optional string: Y axis label}

\item{plt.title}{optional string: Plot title}
}
\value{
Plot of the DCF
}
\description{
Creates a plot of the DCF
}
\examples{
\dontrun{
plt <- dcf.plot(dcf.df)
plot(plt)
}
}
\author{
Derek Blue
}
