% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{ff.plot}
\alias{ff.plot}
\title{Plot Flux Flux}
\usage{
ff.plot(ff.df, xlab = "Soft Rate [count/s]", ylab = "Hard Rate [count/s]")
}
\arguments{
\item{ff.df}{Flux flux data frame with structure: SOFT.RATE, SOFT.ERROR, HARD.RATE, HARD.ERROR}

\item{xlab}{optional: X axis label}

\item{ylab}{optional: Y axis label}
}
\value{
Plot of the hard band flux vs soft band flux
}
\description{
Plots the hard band flux vs soft band flux
}
\examples{
\dontrun{
plt <- ff.plot(ff.df)
plot(plt)
}
}
\author{
Derek Blue
}
