% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{hs.plot}
\alias{hs.plot}
\title{Plot Hard and Soft Light Curves}
\usage{
hs.plot(
  slc,
  hlc,
  xlab = "Time [s]",
  ylab = "Rate [count/s]",
  plt.title = "Hard and Soft Lightcurves"
)
}
\arguments{
\item{slc}{0.3 - 1 keV light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{hlc}{2 - 10 keV light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{xlab}{optional: X axis label}

\item{ylab}{optional: Y axis label}

\item{plt.title}{optional: Plot title}
}
\value{
Plot of the source and background light curve
}
\description{
Plots the source and background of the hard and soft light curves
}
\examples{
\dontrun{
plt <- hs.plot(lightcurve)
plot(plt)
}
}
\author{
Derek Blue
}
