% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{lc.overplot}
\alias{lc.overplot}
\title{Overplot Light Curves}
\usage{
lc.overplot(
  lc1,
  lc2,
  xlab = "Time [s]",
  y1lab = "Rate [count/s]",
  y2lab = "Rate [count/s]",
  plt.title = "LC1 vs LC2",
  bg1 = FALSE,
  bg2 = FALSE,
  names = c("lc1", "lc2")
)
}
\arguments{
\item{lc1}{required data frame: Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{lc2}{required data frame: Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{xlab}{optional string: X axis label}

\item{bg1}{optional bool: If TRUE plot the background of the first light curve, defaults to FALSE}

\item{bg2}{optional bool: If TRUE plot the background of the second light curve, defaults to FALSE}

\item{names}{optional list: Names for the two light curves}

\item{ylab}{optional string: Y axis label}
}
\value{
Plot of the source and background light curve
}
\description{
Overplots the source and background of two light curves
}
\examples{
\dontrun{
plt <- lc.plot(lightcurve)
plot(plt)
}
}
\author{
Derek Blue
}
