% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{lc.plot.error}
\alias{lc.plot.error}
\title{Plot Light Curve Error}
\usage{
lc.plot.error(
  lc,
  bin.width = 0.01,
  xlab = "Rate Error [count/s]",
  ylab = "Density"
)
}
\arguments{
\item{lc}{required data frame: Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{bin.width}{optional numeric: Histogram bin width}

\item{xlab}{optional string: X axis label}

\item{ylab}{optional string: Y axis label}
}
\value{
Plot of the source and background light curve
}
\description{
Creates a plot of the distribution of measurement error for a lightcurve
}
\examples{
\dontrun{
plt <- lc.plot(lightcurve)
plot(plt)
}
}
\author{
Derek Blue
}
