% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{lc.plot}
\alias{lc.plot}
\title{Plot Light Curve}
\usage{
lc.plot(lc, xlab = "Time [s]", ylab = "Rate [count/s]", background = FALSE)
}
\arguments{
\item{lc}{required data frame: Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{xlab}{optional string: X axis label}

\item{ylab}{optional string: Y axis label}

\item{background}{optional bool: If TRUE plot the background of the light curve, defaults to FALSE}
}
\value{
Plot of the source and background light curve
}
\description{
Plots the source and background of a light curve
}
\examples{
\dontrun{
plot(lc.plot(lightcurve, background = TRUE))
}
}
\author{
Derek Blue
}
