% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{lc.plot}
\alias{lc.plot}
\title{Plot Light Curve}
\usage{
lc.plot(
  lc,
  xlab = "Time [s]",
  ylab = "Rate [count/s]",
  plt.title = "Lightcurve",
  background = FALSE,
  color = "black"
)
}
\arguments{
\item{lc}{required data frame: Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}

\item{xlab}{optional string: X axis label}

\item{ylab}{optional string: Y axis label}

\item{plt.title}{optional string: Plot title}

\item{background}{optional boolean: If TRUE plot the background of the light curve, defaults to FALSE}

\item{color}{optional string: plotting color}
}
\value{
Plot of the source and background light curve
}
\description{
Creates a plot of the source and background of a light curve
}
\examples{
\dontrun{
plot(lc.plot(lightcurve, background = TRUE))
}
}
\author{
Derek Blue
}
