% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctools.r
\name{normalize.lc}
\alias{normalize.lc}
\title{Normalize Light Curve}
\usage{
normalize.lc(lc)
}
\arguments{
\item{lc}{required data frame: Light curve data frame with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE}
}
\value{
Light curve data frame normalized by it mean with structure: TIME, TIMED, RATE, ERROR, BACKV, BACKE
}
\description{
Normalizes a light curve based on its mean count rate
}
\examples{
\dontrun{
lc <- xmm.pn.lc(fits_file_path)
lc <- normalize.lc(lc)
}
}
\author{
Derek Blue
}
