% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lctools.r
\name{sim.lc}
\alias{sim.lc}
\title{Simulate Light Curve}
\usage{
sim.lc(beta, bins = 1024, length = 1e+05, scale.factor = 1, shift.factor = 0)
}
\arguments{
\item{beta}{Slope of the PSD for the simulated light curve}

\item{bins}{Number of data points for the simulated light curve, defaults to 1024}

\item{length}{Observation length, in seconds, for the simulated light curve, defaults to 100000}

\item{scale.factor}{Scaling factor for the simulated light curve, defaults to 1}

\item{shift.factor}{Shift factor for the simulated light curve, defaults to 0}
}
\value{
Simulated light curve data frame with structure: TIME, TIMED, RATE
}
\description{
Simulates a light curve from a given power spectrum
}
\examples{
\dontrun{
lightcurve <- sim.lc(lightcurve, 100)
}
}
\author{
Derek Blue
}
